Implement flag constants
The `$flags` argument of `md4c_toHtml()` accepts an int.  However, when
reading code, magic numbers are hard to understand, and the underlying
values may change with new md4c releases.  As such it seems sensible to
actually define respective PHP constants.

We also add a test to check that the `$flags` arguments generally
works, and to also check that the `MD4C_DIALECT_GITHUB` constant does
not include `MD4C_FLAG_NOINDENTEDCODEBLOCKS` (which is contained in the
default value).